+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - Speed up np.ndarray(..., dtype=np.float64) @ MatrixExpr
 - MatrixExpr and MatrixExprCons use `__array_ufunc__` protocol to control all numpy.ufunc inputs and outputs
 - Set `__array_priority__` for MatrixExpr and MatrixExprCons
+- changed addConsNode() and addConsLocal() to mirror addCons() and accept ExprCons instead of Constraint
 ### Removed
 
 ## 6.0.0 - 2025.xx.yy